LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-fr-CA.html

<!DOCTYPE html><!-- webkit-test-runner [ language=fr-CA ] -->
<html>
<head>
<script src="../../../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
shouldBeEqualToString("navigator.language", "fr-ca");
</script>
<script src="../../../../../resources/js-test-post.js"></script>
</body>
</html>
