LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-fr-CA-expected.txt

PASS navigator.language is "fr-CA"
PASS successfullyParsed is true

TEST COMPLETE
